Refactor to have directionality
